serialio: Support for mmap serial ports
Some chipsets have memory mapped serial ports. The protocol is the same
as an standard uart, but with memory read/write instead of inb/outb.

Message-Id: <20161220180358.5567-1-ricardo.ribalda@gmail.com>
Tested-by: Ricardo Ribalda Delgado <ricardo.ribalda@gmail.com>

+++ src/serial.c
@@ -21,7 +21,7 @@
 static u16
 detect_serial(u16 port, u8 timeout, u8 count)
 {
-    if (CONFIG_DEBUG_SERIAL && port == CONFIG_DEBUG_SERIAL_PORT
+    if (CONFIG_DEBUG_SERIAL && port == (u16) CONFIG_DEBUG_SERIAL_PORT
         && !romfile_loadint("etc/advertise-serial-debug-port", 1))
         return 0;
     outb(0x02, port+SEROFF_IER);
